Added reuseable HoverableTextfield component
